adding AutorMapper, LivroMapper and updating ExceptionHandle

+++ src/main/java/com/github/iuryrayam/libraryapi/service/AutorService.java
@@ -1,7 +1,13 @@
 package com.github.iuryrayam.libraryapi.service;
 
+import com.github.iuryrayam.libraryapi.exception.OperacaoNaoPermitidaException;
 import com.github.iuryrayam.libraryapi.model.Autor;
 import com.github.iuryrayam.libraryapi.repository.AutorRepository;
+import com.github.iuryrayam.libraryapi.repository.LivroRepository;
+import com.github.iuryrayam.libraryapi.validator.AutorValidator;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Example;
+import org.springframework.data.domain.ExampleMatcher;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -9,15 +15,15 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor
 public class AutorService {
 
-    private AutorRepository repository;
-
-    public AutorService(AutorRepository repository){
-        this.repository = repository;
-    }
+    private final AutorRepository repository;
+    private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
     public Autor salvar(Autor autor){
+        validator.validar(autor);
         return repository.save(autor);
     }
 
@@ -26,6 +32,10 @@ public class AutorService {
     }
 
     public void delete(Autor autor){
+        if (possuiLivro(autor)){
+            throw new OperacaoNaoPermitidaException(
+                    "Não é permitido excluir um autor que possui livros cadastrados!");
+        }
         repository.delete(autor);
     }
 
@@ -45,10 +55,30 @@ public class AutorService {
         return repository.findAll();
     }
 
+    public List<Autor> pesquisaByExample(String nome, String nacionalidade){
+        Autor autor = new Autor();
+        autor.setNome(nome);
+        autor.setNacionalidade(nacionalidade);
+
+        ExampleMatcher matcher = ExampleMatcher.matching()
+                .withIgnorePaths("id", "data_nascimento", "data_cadastro")
+                .withIgnoreNullValues()
+                .withIgnoreCase()
+                .withStringMatcher(ExampleMatcher.StringMatcher.CONTAINING);
+
+        Example<Autor> autorExample = Example.of(autor, matcher);
+        return repository.findAll(autorExample);
+    }
+
     public void atualizar(Autor autor){
         if (autor.getId() == null){
             throw new IllegalArgumentException("Para atualizar, é necessário que o autor ja esteja salvo na base.");
         }
+        validator.validar(autor);
         repository.save(autor);
+    }
+
+    private boolean possuiLivro(Autor autor){
+        return livroRepository.existsByAutor(autor);
     }
 }
